pyproject.toml
[tool.poetry]
name        = "llamascholar"
version     = "0.0.1"
description = "Research-assistant agent powered by Llama-3 on Workers AI"
authors     = ["Karthik Alladi <kalladi@usc.edu>"]

packages = [{ include = "llamascholar" }]

# —— Runtime requirements ————————————————————————————————
[tool.poetry.dependencies]
python                 = "^3.11"

# core stack
duckduckgo-search      = "^8.0.4"
langchain-cloudflare   = "^0.1.8"
langchain-openai       = "^0.3.26"
arxiv                  = "^2.2.0"

# —— Development / test extras ————————————————————————————
python-dotenv = "^1.1.1"
langchain = "^0.3.26"
[tool.poetry.group.dev.dependencies]
pytest     = "^8.2"
black      = "^24.4"

# —— Tool configuration examples (optional) ——————————————
[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
addopts = "-q"

# —— Build backend footer ————————————————————————————————
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
